Adds detail.html and list.html templates
Template list.html allow to list all active users in the site with
profile images thumbnails. Detail template display the user profile and
the thumbnail profile image. Shows the total number of followers and a link
to follow/unfollow. Images are display by the user with the list_ajax.html
template